src/sim-engine/requirements.txt
google-genai
python-dotenv
langchain[google-genai]
langgraph
